chore: Entity 패키지명 수정
- 패키지 이름 규칙에 맞게 Entity의 첫글자를 소문자로 변경
