+++ src/main/java/com/example/point_career/domain/bookmark/entity/PointBookmark.java
@@ -1,13 +1,14 @@
-package com.example.point_career.domain.bookmark.Entity;
+package com.example.point_career.domain.bookmark.entity;
 
-import com.example.point_career.domain.recruit.Entity.Recruit;
-import com.example.point_career.domain.user.Entity.User;
+import com.example.point_career.domain.point.entity.Point;
+import com.example.point_career.domain.user.entity.User;
 import jakarta.persistence.*;
+
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "RecruitBookmark")
-public class RecruitBookmark {
+@Table(name = "PointBookmark")
+public class PointBookmark {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id; // 북마크 ID, 기본키
 
@@ -18,8 +19,9 @@ public class RecruitBookmark {
     @JoinColumn(name = "user_id")
     private User user;
 
+    // 포인트 활동 북마크
     @ManyToOne
-    @JoinColumn(name = "recruit_id")
-    private Recruit recruit;
-}
+    @JoinColumn(name = "point_id")
+    private Point point;
 
+}
