+++ src/main/java/com/example/point_career/domain/bookmark/entity/RecruitBookmark.java
@@ -1,14 +1,13 @@
-package com.example.point_career.domain.bookmark.Entity;
+package com.example.point_career.domain.bookmark.entity;
 
-import com.example.point_career.domain.point.Entity.Point;
-import com.example.point_career.domain.user.Entity.User;
+import com.example.point_career.domain.recruit.entity.Recruit;
+import com.example.point_career.domain.user.entity.User;
 import jakarta.persistence.*;
-
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "PointBookmark")
-public class PointBookmark {
+@Table(name = "RecruitBookmark")
+public class RecruitBookmark {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id; // 북마크 ID, 기본키
 
@@ -19,9 +18,8 @@ public class PointBookmark {
     @JoinColumn(name = "user_id")
     private User user;
 
-    // 포인트 활동 북마크
     @ManyToOne
-    @JoinColumn(name = "point_id")
-    private Point point;
-
+    @JoinColumn(name = "recruit_id")
+    private Recruit recruit;
 }
+
