--- src/main/java/com/example/point_career/domain/tag/Entity/TagType.java
@@ -1,5 +1,0 @@
-package com.example.point_career.domain.tag.Entity;
-
-public enum TagType {
-	INDUSTRY, FAVORITE
-}
