+++ src/main/java/com/example/point_career/domain/tag/entity/PointTag.java
@@ -1,6 +1,6 @@
-package com.example.point_career.domain.tag.Entity;
+package com.example.point_career.domain.tag.entity;
 
-import com.example.point_career.domain.point.Entity.Point;
+import com.example.point_career.domain.point.entity.Point;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
