+++ src/main/java/com/example/point_career/domain/tag/entity/TagType.java
@@ -1,0 +1,5 @@
+package com.example.point_career.domain.tag.entity;
+
+public enum TagType {
+	INDUSTRY, FAVORITE
+}
